+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Slim-Lint Changelog
 
+# master (unreleased)
+
+* Add support for `include`/`exclude` options on linters, allowing a list of
+  files or glob patterns to be included/excluded from a linter's scope
+
 ## 0.2.0
 
 * Ignore `Style/IndentationConsistency` RuboCop warnings